use duration for timer

+++ app/src/main/java/com/fesvieira/pomotasks/helpers/formatToString.kt
@@ -1,5 +1,5 @@
 package com.fesvieira.pomotasks.helpers
 
-val Int.formatToString get(): String {
+val Long.formatToString get(): String {
     return if (this > 9) "$this" else "0$this"
 }
